backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
fpdf2
python-multipart
nvdlib
requests
